+++ test/AutoCompleteSpec.js
@@ -2,39 +2,38 @@ import React from 'react';
 import { findDOMNode } from 'react-dom';
 import ReactTestUtils from 'react-dom/test-utils';
 import AutoComplete from '../src/AutoComplete';
+import { getDOMNode, getInstance } from './TestWrapper';
 
 describe('AutoComplete', () => {
   it('Should render input', () => {
-    const instance = ReactTestUtils.renderIntoDocument(<AutoComplete />);
+    const instance = getInstance(<AutoComplete />);
 
     assert.ok(ReactTestUtils.findRenderedDOMComponentWithTag(instance, 'input'));
   });
 
   it('Should render 2 `li` when set `open` and `defaultValue`', () => {
-    const instance = ReactTestUtils.renderIntoDocument(
-      <AutoComplete data={['a', 'b', 'ab']} open defaultValue="a" />
-    );
+    const instance = getInstance(<AutoComplete data={['a', 'b', 'ab']} open defaultValue="a" />);
     assert.equal(instance.menuContainer.querySelectorAll('li').length, 2);
   });
 
   it('Should be a `top-right` for placement', () => {
-    const instance = ReactTestUtils.renderIntoDocument(<AutoComplete open placement="topRight" />);
+    const instance = getInstance(<AutoComplete open placement="topRight" />);
     const classes = instance.menuContainer.parentNode.className;
     assert.include(classes, 'rs-placement-top-right');
   });
 
   it('Should be disabled', () => {
-    const instance = ReactTestUtils.renderIntoDocument(<AutoComplete disabled />);
-    assert.include(findDOMNode(instance).className, 'rs-auto-complete-disabled');
+    const instance = getDOMNode(<AutoComplete disabled />);
+    assert.include(instance.className, 'rs-auto-complete-disabled');
   });
 
   it('Should call onSelect callback', done => {
-    const doneOp = value => {
-      if (value === 'a') {
+    const doneOp = item => {
+      if (item.value === 'a') {
         done();
       }
     };
-    const instance = ReactTestUtils.renderIntoDocument(
+    const instance = getInstance(
       <AutoComplete data={['a', 'b', 'ab']} open defaultValue="a" onSelect={doneOp} />
     );
     ReactTestUtils.Simulate.click(instance.menuContainer.querySelectorAll('a')[0]);
@@ -44,8 +43,8 @@ describe('AutoComplete', () => {
     const doneOp = () => {
       done();
     };
-    const instance = ReactTestUtils.renderIntoDocument(<AutoComplete onChange={doneOp} />);
-    const input = findDOMNode(instance).querySelector('input');
+    const instance = getDOMNode(<AutoComplete onChange={doneOp} />);
+    const input = instance.querySelector('input');
 
     input.value = 'a';
 
@@ -56,8 +55,8 @@ describe('AutoComplete', () => {
     const doneOp = () => {
       done();
     };
-    const instance = ReactTestUtils.renderIntoDocument(<AutoComplete onFocus={doneOp} />);
-    const input = findDOMNode(instance).querySelector('input');
+    const instance = getDOMNode(<AutoComplete onFocus={doneOp} />);
+    const input = instance.querySelector('input');
     ReactTestUtils.Simulate.focus(input);
   });
 
@@ -65,13 +64,13 @@ describe('AutoComplete', () => {
     const doneOp = () => {
       done();
     };
-    const instance = ReactTestUtils.renderIntoDocument(<AutoComplete onBlur={doneOp} />);
-    const input = findDOMNode(instance).querySelector('input');
+    const instance = getDOMNode(<AutoComplete onBlur={doneOp} />);
+    const input = instance.querySelector('input');
     ReactTestUtils.Simulate.blur(input);
   });
 
   it('Should render a icon in li', () => {
-    const instance = ReactTestUtils.renderIntoDocument(
+    const instance = getInstance(
       <AutoComplete
         data={['a', 'b', 'ab']}
         open
@@ -84,13 +83,13 @@ describe('AutoComplete', () => {
   });
 
   it('Should have a custom className', () => {
-    const instance = ReactTestUtils.renderIntoDocument(<AutoComplete className="custom" />);
-    assert.include(findDOMNode(instance).className, 'custom');
+    const instance = getDOMNode(<AutoComplete className="custom" />);
+    assert.include(instance.className, 'custom');
   });
 
   it('Should have a custom style', () => {
     const fontSize = '12px';
-    const instance = ReactTestUtils.renderIntoDocument(<AutoComplete style={{ fontSize }} />);
-    assert.equal(findDOMNode(instance).style.fontSize, fontSize);
+    const instance = getDOMNode(<AutoComplete style={{ fontSize }} />);
+    assert.equal(instance.style.fontSize, fontSize);
   });
 });